refactor: 컴포넌트 이름 및 구조 변경
