+++ Rolling/src/App.jsx
@@ -1,9 +1,9 @@
-import PostPage from './pages/PostPage';
+import MessagePage from './pages/MessagePage';
 
 function App() {
   return (
     <>
-      <PostPage />
+      <MessagePage />
       {/* useLoaction을 사용하여 PostEditPage를 보여주게 할 지 미정 */}
     </>
   );
